# swagger/swagger-output.yaml
openapi: 3.0.0
info:
  title: School Management API
  version: 1.0.0
  description: API for school admin storage and management
servers:
  - url: https://cse-341-group-project-mfvd.onrender.com/api
    description: Production server
  - url: http://localhost:5500/api
    description: Local development server
components:
  securitySchemes:
    googleOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID access
            profile: View your profile info
            email: View your email address
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    class:
      type: object
      properties:
        _id:
          type: string
          readOnly: true
        name:
          type: string
          example: SS1 A
        section:
          type: string
          example: Science
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - name
        - section
    department:
      type: object
      properties:
        _id:
          type: string
          readOnly: true
        name:
          type: string
          example: Mathematics
        description:
          type: string
          example: Handles all mathematics courses
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - name
    course:
      type: object
      properties:
        _id:
          type: string
          readOnly: true
        title:
          type: string
          example: Algebra 101
        description:
          type: string
          example: Introductory algebra course
        department:
          type: string
          example: 60f6e9f2b7a1c142d8ef8c11
        instructor:
          type: string
          example: 60f6e9f2b7a1c142d8ef8c12
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - title
        - department
        - instructor
    enrollment:
      type: object
      properties:
        _id:
          type: string
          readOnly: true
        student:
          type: string
          example: 60f6e9f2b7a1c142d8ef8c10
        course:
          type: string
          example: 60f6e9f2b7a1c142d8ef8c11
        enrolledAt:
          type: string
          format: date-time
          example: '2025-01-10T12:00:00Z'
        grade:
          type: string
          enum:
            - A
            - B
            - C
            - D
            - E
            - F
            - I
            - W
            - null
          example: A
        status:
          type: string
          enum:
            - enrolled
            - completed
            - withdrawn
          example: enrolled
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - student
        - course
security:
  - bearerAuth: []
paths:
...
  /auth/logout:
    get:
      summary: Logout user
      tags:
        - Authentication
      description: Clears the session and invalidates the JWT token
      responses:
        '302':
          description: Redirect to success page
        '401':
          description: User not authenticated
  /user/protected:
    get:
      summary: Test protected route with JWT
      security:
        - bearerAuth: []
      tags:
        - User
      responses:
        '200':
          description: Authorized user access
        '401':
          description: Unauthorized
  /class:
    post:
      summary: Create a new class
      tags:
        - Class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/class'
      responses:
        '201':
          description: Class created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/class'
        '400':
          description: Invalid input
    get:
      summary: Get all classes
      tags:
        - Class
      responses:
        '200':
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/class'
  /class/{id}:
    get:
      summary: Get a class by ID
      tags:
        - Class
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The class ID
      responses:
        '200':
          description: Class found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/class'
        '404':
          description: Class not found
    put:
      summary: Update a class by ID
      tags:
        - Class
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The class ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/class'
      responses:
        '200':
          description: Class updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/class'
        '400':
          description: Invalid input
        '404':
          description: Class not found
    delete:
      summary: Delete a class member by ID
      tags:
        - Class
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The class ID
      responses:
        '200':
          description: Class member deleted
        '404':
          description: Class not found
  /course:
    post:
      summary: Create a new course
      tags:
        - Course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/course'
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/course'
        '400':
          description: Invalid input
    get:
      summary: Get all courses
      tags:
        - Course
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/course'
  /course/{id}:
    get:
      summary: Get a course by ID
      tags:
        - Course
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The course ID
      responses:
        '200':
          description: Course found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/course'
        '404':
          description: Course not found
    put:
      summary: Update a course by ID
      tags:
        - Course
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The course ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/course'
      responses:
        '200':
          description: Course updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/course'
        '400':
          description: Invalid input
        '404':
          description: Course not found
    delete:
      summary: Delete a course by ID
      tags:
        - Course
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The course ID
      responses:
        '204':
          description: Course deleted successfully
        '404':
          description: Course not found
  /department:
    post:
      summary: Create a new department
      tags:
        - Department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/department'
      responses:
        '201':
          description: Department created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/department'
        '400':
          description: Invalid input
    get:
      summary: Get all department
      tags:
        - Department
      responses:
        '200':
          description: List of department
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/department'
  /dpeartment/{id}:
    get:
      summary: Get a department by ID
      tags:
        - Department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The department ID
      responses:
        '200':
          description: Department found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/department'
        '404':
          description: User not found
  /deparment/{id}:
    put:
      summary: Update a deparment by ID
      tags:
        - Department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The department ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/department'
      responses:
        '200':
          description: Department updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/department'
        '400':
          description: Invalid input
        '404':
          description: Department not found
  /department/{id}:
    delete:
      summary: Delete a department member by ID
      tags:
        - Department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The department ID
      responses:
        '200':
          description: Department member deleted
        '404':
          description: Department not found
...
  /enrollment/course/{courseId}:
    get:
      summary: Get enrollments for a specific course
      tags:
        - Enrollment
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
          description: The course ID
      responses:
        '200':
          description: List of course enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/enrollment'
        '404':
          description: Course not found
...